Refactor BOOT out of dos_programs
